Move `review_prefs` to a separate module
